+++ tests/mutation/test_mutation_runtime.py
@@ -20,7 +20,8 @@ class Color(Enum):
         return self.name.lower()
 """.strip()
 
-    mutated_code, mutant_names, _ = mutate_file_contents("test.py", source)
+    mutated_file = mutate_file_contents("test.py", source)
+    mutated_code, mutant_names = mutated_file.code, mutated_file.mutant_names
     assert len(mutant_names) > 0, "Should have at least one mutant"
 
     monkeypatch.setenv("MUTANT_UNDER_TEST", "none")
@@ -65,7 +66,8 @@ class Color(Enum):
         return vals[name]
 """.strip()
 
-    mutated_code, mutant_names, _ = mutate_file_contents("test.py", source)
+    mutated_file = mutate_file_contents("test.py", source)
+    mutated_code, mutant_names = mutated_file.code, mutated_file.mutant_names
     assert len(mutant_names) > 0, "Should have at least one mutant"
 
     monkeypatch.setenv("MUTANT_UNDER_TEST", "none")
@@ -88,7 +90,8 @@ class Calculator:
         return a + b
 """.strip()
 
-    mutated_code, mutant_names, _ = mutate_file_contents("test.py", source)
+    mutated_file = mutate_file_contents("test.py", source)
+    mutated_code, mutant_names = mutated_file.code, mutated_file.mutant_names
     assert len(mutant_names) > 0, "Should have at least one mutant"
 
     monkeypatch.setenv("MUTANT_UNDER_TEST", "none")
@@ -117,7 +120,8 @@ class SubClass(Base):
         self.value = value
 """.strip()
 
-    mutated_code, mutant_names, _ = mutate_file_contents("test.py", source)
+    mutated_file = mutate_file_contents("test.py", source)
+    mutated_code, mutant_names = mutated_file.code, mutated_file.mutant_names
     assert len(mutant_names) > 0, "Should have at least one mutant"
 
     monkeypatch.setenv("MUTANT_UNDER_TEST", "none")
@@ -140,7 +144,8 @@ def foo(a: int, b: int = 2):
     return a + b
 """.strip()
 
-    mutated_code, mutant_names, _ = mutate_file_contents("test.py", source)
+    mutated_file = mutate_file_contents("test.py", source)
+    mutated_code, mutant_names = mutated_file.code, mutated_file.mutant_names
     assert len(mutant_names) > 0, "Should have at least one mutant"
 
     monkeypatch.setenv("MUTANT_UNDER_TEST", "none")